requirements.txt
flask==3.0.0
groq==0.9.0
python-dotenv==1.0.0
requests==2.31.0
